fix: Corrected Scheme Composition handling approach
