+++ .github/workflows/npm-publish.yaml
@@ -1,0 +1,35 @@
+name: Publish master-exploder npm package
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  publish:
+    needs: workspace-changes
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 18
+          cache: yarn
+
+      - name: Install dependencies
+        run: yarn install
+
+      - name: Configure npm with token
+        run: |
+          npm config set @api-stuff:registry https://npm.pkg.github.com
+          npm config set //npm.pkg.github.com/:_authToken=${NPM_TOKEN}
+        env:
+          NPM_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Publish package to GitHub Package Registry
+        run: yarn workspace @api-stuff/master-exploder publish --non-interactive
+        env:
+          NPM_TOKEN: ${{ secrets.GITHUB_TOKEN }}
