+++ .github/workflows/version-update.yaml
@@ -1,0 +1,37 @@
+name: Version Bump in Pull Request
+
+on:
+  pull_request:
+    branches:
+      - develop
+
+jobs:
+  bump-version:
+    runs-on: ubuntu-latest
+    needs: workspace-changes
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          ref: ${{ github.event.pull_request.head.ref }}
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 18
+          cache: yarn
+
+      - name: Install dependencies
+        run: yarn install
+
+      - name: Configure Git user
+        run: |
+          git config --global user.name "GitHub Actions Bot"
+          git config --global user.email "actions@github.com"
+
+      - name: Bump API design ruleset version and commit
+        run: |
+          yarn version --patch --no-git-tag-version
+          git add package.json
+          git commit -m "Auto bump version to $(jq -r .version package.json)"
+          git push origin ${{ github.event.pull_request.head.ref }}
