+++ lib/schema-object/index.js
@@ -1,136 +1,18 @@
-const { getJsonPath } = require('./util');
+const {
+  getParentArrayPath,
+  getTypeFunction,
+} = require('./type-functions');
+const { getJsonPath } = require('../util');
 
 const keywords = ['items', 'type', 'properties', 'required', 'additionalProperties', 'minLength', 'maxLength', 'minItems', 'maxItems', 'minProperties', 'maxProperties', 'enum', 'format', 'title', 'pattern', 'allOf', 'oneOf', 'anyOf', 'not'];
+
+const isSchemaCompositionObject = (path) => /(allOf|anyOf|not|oneOf)\.[0-9]+$/.test(path.join('.'));
 
 const isKeywordExtensionOrIndex = (fragment) => keywords
   .indexOf(fragment) !== -1 || Number.isInteger(parseInt(fragment)) || fragment.match(/x-.*/);
 
 const getSerialisationPath = (path) => path
   .filter((fragment) => !isKeywordExtensionOrIndex(fragment));
-
-const getParentArrayPath = (path, arrays) => Object.entries(arrays)
-  .filter(([key, value]) => path.startsWith(key)) // eslint-disable-line no-unused-vars
-  .map(([source, replacement]) => ({ source, replacement }))
-  .pop();
-
-const normaliseDescription = (description) => (description || '').replace(/\n|\s{2,}/g, ' ').replace(/ +/g, ' ');
-
-const getArrayRequirement = (dataPath, isRequired, child, arrays) => {
-  const coreRequirement = {
-    type: () => 'array',
-    required: () => isRequired,
-    minItems: () => child.value.minItems || '',
-    maxItems: () => child.value.maxItems || '',
-    additionalProperties: () => 'N/A',
-    description: () => normaliseDescription(child.value.description),
-  };
-
-  // Get any parent arrays and update path with parent
-  const parentArray = getParentArrayPath(dataPath, arrays);
-  const rawArrayPath = parentArray
-    ? dataPath.replace(parentArray.source, parentArray.replacement)
-    : dataPath;
-
-  const arrayPath = rawArrayPath.match(/\[0\]$/) ? rawArrayPath : `${rawArrayPath}[*]`;
-  arrays[dataPath.replace(/\[[0-9]+\]/, '')] = arrayPath; // eslint-disable-line no-param-reassign
-
-  return {
-    coreRequirement,
-    ...(child.value.items.type !== 'object' ? {
-      additionalRequirement: {
-        additionalDataPath: arrayPath,
-        value: child.value.items,
-      },
-    } : {}),
-  };
-};
-
-const getObjectRequirement = (
-  dataPath,
-  isRequired,
-  child,
-) => {
-  const { additionalProperties, description } = child.value;
-
-  return {
-    type: () => 'object',
-    required: () => isRequired,
-    additionalProperties: () => additionalProperties !== false,
-    description: () => normaliseDescription(description),
-  };
-};
-
-const getStringRequirement = (
-  dataPath,
-  isRequired,
-  child,
-) => ({
-  type: () => 'string',
-  required: () => isRequired,
-  minLength: () => child.value.minLength || '',
-  maxLength: () => child.value.maxLength || '',
-  format: () => child.value.format || '',
-  pattern: () => child.value.pattern || '',
-  enum: () => (child.value.enum || []).join(', '),
-  additionalProperties: () => 'N/A',
-  description: () => normaliseDescription(child.value.description),
-});
-
-const getIntegerRequirement = (
-  dataPath,
-  isRequired,
-  child,
-) => ({
-  type: () => 'integer',
-  required: () => isRequired,
-  format: () => child.value.format || '',
-  additionalProperties: () => 'N/A',
-  description: () => normaliseDescription(child.value.description),
-});
-
-const getNumberRequirement = (
-  dataPath,
-  isRequired,
-  child,
-) => ({
-  type: () => 'number',
-  required: () => isRequired,
-  format: () => child.value.format || '',
-  additionalProperties: () => 'N/A',
-  description: () => normaliseDescription(child.value.description),
-});
-
-// eslint-disable-next-line no-unused-vars
-const getBooleanRequirement = (
-  dataPath,
-  isRequired,
-  child,
-) => ({
-  type: () => 'boolean',
-  required: () => isRequired,
-  additionalProperties: () => 'N/A',
-  description: () => normaliseDescription(child.value.description),
-});
-
-const getTypeFunction = (type, definition, requirementType) => {
-  const requirementFunction = {
-    array: getArrayRequirement,
-    boolean: getBooleanRequirement,
-    integer: getIntegerRequirement,
-    number: getNumberRequirement,
-    object: getObjectRequirement,
-    string: getStringRequirement,
-  };
-
-  const fn = requirementFunction[type];
-
-  if (!fn) {
-    console.error(definition);
-    throw new Error(`Could not match type value to transform function ${type} in ${requirementType} requirement`);
-  }
-
-  return fn;
-};
 
 class SchemaObject {
   constructor(schemaObject, delimiter) {
@@ -194,8 +76,15 @@ class SchemaObject {
 
     return children
       .reduce((output, child) => {
-        // Only process items that can actually be seriailised i.e a named schema object
-        if (child.value.type && !isKeywordExtensionOrIndex(child.path.slice(-1).pop())) {
+        // Only process items that can actually be serialised i.e a named schema object
+        // In this content the type property must exist and:
+        //
+        // * The property is nested in a Schema Composition object
+        // * This is not a keyword or index (this is belt and braces and could possibly be removed)
+        //
+        if (child.value.type
+          && (isSchemaCompositionObject(child.path)
+            || !isKeywordExtensionOrIndex(child.path.slice(-1).pop()))) {
           const { path } = this.schemaObject;
           const [, , url, method, , contentPath] = path || [];
           const pathOrSchemaObjectName = url || this.schemaObjectName;
@@ -268,7 +157,7 @@ class SchemaObject {
 
           if (!schemaComposition && !value.type) {
             console.error(value);
-            throw new Error('Not imeplemented');
+            throw new Error('Not implemented');
           }
 
           // Catch-all loop over additional requirements
@@ -276,15 +165,15 @@ class SchemaObject {
             value.type
               ? [value] // Has a type set
               : schemaComposition) // Has a composition keyword set
-            .map((additionaPropertyValue) => {
-              const additionalRequirementFn = getTypeFunction(additionaPropertyValue.type, additionaPropertyValue, 'additional');
+            .map((additionalPropertyValue) => {
+              const additionalRequirementFn = getTypeFunction(additionalPropertyValue.type, additionalPropertyValue, 'additional');
 
               return this.getOutputValues(
                 additionalRequirementResponseSignature,
                 additionalRequirementFn(
                   additionalDataPath,
                   isRequired,
-                  { value: additionaPropertyValue },
+                  { value: additionalPropertyValue },
                   this.arrays,
                 ),
               );
